Moved user query to core
