+++ app/controllers/users.js
@@ -36,24 +36,7 @@ module.exports = function() {
                     take: req.param('take')
                 };
 
-            options = helpers.sanitizeQuery(options, {
-                defaults: {
-                    take: 500
-                },
-                maxTake: 5000
-            });
-
-            var find = User.find();
-
-            if (options.skip) {
-                find.skip(options.skip);
-            }
-
-            if (options.take) {
-                find.limit(options.take);
-            }
-
-            find.exec(function(err, users) {
+            core.user.list(options, function(err, users) {
                 if (err) {
                     console.log(err);
                     return res.status(400).json(err);
